tests: locpick, beacon v0.1, v0.2 tests
