--- src/beacon/.gitignore
@@ -1,3 +1,0 @@
-beacon-devshell-ss-hlsvc.yaml
-beacon-za-ss-hlsvc.yaml
-beacon-zb-ss-hlsvc.yaml
