+++ src/beacon/beacon-msvc/app/beacon.test.js
