+++ src/beacon/beacon-msvc/app/package.json
@@ -1,28 +1,30 @@
 {
-  "name": "beatrak-beacon",
-  "version": "0.0.1",
-  "description": "Node.js on Docker",
-  "author": "Dmitri Chtchourov <dchtchou@cisco.com>",
-  "main": "server.js",
-  "repository": "github.com/cdmitri/cio",
-  "license": "MIT",
-  "private": true,
-  "jest": {
-    "testEnvironment": "node"
-  },
-  "scripts": {
-    "start": "node server.js",
-    "test": "jest --watchAll",
-    "dev": "/start-dev.sh"
-  },
-  "dependencies": {
-    "axios": "^0.17.1",
-    "elasticsearch": "^13.3.1",
-    "express": "^4.13.3",
-    "jest": "^21.2.1",
-    "loglevel": "^1.5.1",
-    "moment": "^2.19.1",
-    "nocache": "^2.0.0",
-    "system-sleep": "^1.3.6"
-  }
+    "name": "beatrak-beacon",
+    "version": "0.0.1",
+    "description": "Node.js on Docker",
+    "author": "Dmitri Chtchourov <dchtchou@cisco.com>",
+    "main": "server.js",
+    "repository": "github.com/cdmitri/cio",
+    "license": "MIT",
+    "private": true,
+    "jest": {
+	"testEnvironment": "node"
+    },
+    "scripts": {
+	"start": "node server.js",
+	"test": "jest --watchAll",
+	"dev": "/start-dev.sh"
+    },
+    "dependencies": {
+	"axios": "^0.17.1",
+	"elasticsearch": "^13.3.1",
+	"express": "^4.13.3",
+	"grpc": "1.10.1",
+	"jest": "^21.2.1",
+	"loglevel": "^1.5.1",
+	"moment": "^2.19.1",
+	"nocache": "^2.0.0",
+	"system-sleep": "^1.3.6",
+	"node-getopt": "^0.3.2"
+    }
 }
