+++ test/TEST.md
@@ -2,23 +2,38 @@
 [user@beatrak beatrak]$ cd test/api
 [user@beatrak api]$ make test-prereq
 [user@beatrak api]$ make all
+[cdmitri@beatrak api]$ make
 ./test-api.sh
-TESTS=.
-/home/cdmitri/src/github.com/ciscosystems/beatrak/test/api
-/home/cdmitri/src/github.com/ciscosystems/beatrak/test/api
-++ time bats --tap .
-1..8
-ok 1 test-beacon-locpick-run-fail-v0.1
-ok 2 test-locpick-run-v0.1
-ok 3 test-locpick-get-info-v0.1
-ok 4 test-locpick-get-locs-v0.1
-ok 5 test-locpick-get-loc-nozone-v0.1
-ok 6 test-locpick-get-loc-zone-v0.1
-ok 7 test-locpick-reset-v0.1
-ok 8 test-locpick-get-config-v0.1
+/home/cdmitri/src/github.com/cdmitri/beatrak/test/api
+++ time bats --tap test-api-locpick.bats
+1..7
+ok 1 test-locpick-run
+ok 2 test-locpick-http-get-info
+ok 3 test-locpick-http-get-locs
+ok 4 test-locpick-http-gen-nozone
+ok 5 test-locpick-http-gen-loc-zone
+ok 6 test-locpick-http-reset
+ok 7 test-locpick-http-get-config
 
-real	0m7.359s
-user	0m0.890s
-sys	0m0.204s
+real	0m8.008s
+user	0m0.611s
+sys	0m0.171s
+/home/cdmitri/src/github.com/cdmitri/beatrak/test/api
+/home/cdmitri/src/github.com/cdmitri/beatrak/test/api
+++ time bats --tap test-api-beacon.bats
+1..9
+ok 1 test-beacon-locpick-run
+ok 2 test-beacon-run
+ok 3 test-beacon-locpick-http-info
+ok 4 test-beacon-locpick-grpc-info
+ok 5 test-beacon-locpick-grpc-tls-info
+ok 6 test-beacon-http-server-init
+ok 7 test-beacon-grpc-init
+ok 8 test-beacon-get-loc
+ok 9 test-beacon-init-done
+
+real	0m10.103s
+user	0m4.263s
+sys	0m0.294s
 ```
